cleanup unused files for data connections
