+++ public/components/app.tsx
@@ -19,7 +19,7 @@ import { EventAnalytics } from './event_analytics';
 import { Home as MetricsHome } from './metrics/index';
 import { Main as NotebooksHome } from './notebooks/components/main';
 import { Home as TraceAnalyticsHome } from './trace_analytics/home';
-import { Home as DatasourcesHome } from './flint/home';
+import { Home as DatasourcesHome } from './data_connections/home';
 
 interface ObservabilityAppDeps {
   CoreStartProp: CoreStart;
