+++ public/components/data_connections/components/datasource.tsx
@@ -24,6 +24,13 @@ import {
   EuiInMemoryTable,
   EuiTableFieldDataColumnType,
   EuiPageContent,
+  EuiPanel,
+  EuiPageContentHeaderSection,
+  EuiPageHeader,
+  EuiPageHeaderSection,
+  EuiAccordion,
+  EuiIcon,
+  EuiCard,
 } from '@elastic/eui';
 import React, { ReactChild, useEffect, useState } from 'react';
 import httpClientMock from 'test/__mocks__/httpClientMock';
@@ -34,17 +41,18 @@ import {
   OPENSEARCH_DOCUMENTATION_URL,
 } from '../../../../common/constants/integrations';
 import { AccelerateFlyout } from './accelerate_flyout';
+import { TabbedPage } from '../../common/tabbed_page/tabbed_page';
 
-export function Table(props: any) {
+export function DataSource(props: any) {
   // const [isFlyoutVisible, setIsFlyoutVisible] = useState(false);
-  const { table, dataSource, pplService } = props;
-  console.log(table);
+  const { dataSource, pplService } = props;
+  console.log(dataSource);
 
   const [tables, setTables] = useState([]);
 
   useEffect(() => {
     pplService
-      .fetch({ query: `source = ${dataSource}.sql(\'DESCRIBE TABLE ${table}\')`, format: 'jdbc' })
+      .fetch({ query: `source = ${dataSource}.sql(\'SHOW TABLES\')`, format: 'jdbc' })
       .then((data) =>
         setTables(
           data.jsonData.map((x: any) => {
@@ -114,25 +122,54 @@ export function Table(props: any) {
 
   return (
     <EuiPage>
-      <EuiPageContent data-test-subj="addedIntegrationsArea">
-        {dataSource}
-        {/* <EuiSpacer/>
+      <EuiPageBody>
+        <EuiPageHeader style={{ justifyContent: 'spaceBetween' }}>
+          <EuiPageHeaderSection style={{ width: '100%', justifyContent: 'space-between' }}>
+            <EuiFlexGroup>
+              <EuiFlexItem grow={false}>
+                <EuiTitle data-test-subj="eventHomePageTitle" size="l">
+                  <h1>{dataSource}</h1>
+                </EuiTitle>
+              </EuiFlexItem>
+            </EuiFlexGroup>
+          </EuiPageHeaderSection>
+        </EuiPageHeader>
 
-      <EuiInMemoryTable
-          loading={props.loading}
-          items={tables}
-          itemId="id"
-          columns={tableColumns}
-          tableLayout="auto"
-          pagination={{
-            initialPageSize: 10,
-            pageSizeOptions: [5, 10, 15],
-          }}
-          search={search}
-          allowNeutralSort={false}
-          isSelectable={true}
-        /> */}
-      </EuiPageContent>
+        <EuiPanel>{dataSource}</EuiPanel>
+        <EuiSpacer />
+        <EuiAccordion
+          id="queryOrAccelerateAccordion"
+          buttonContent="Ways to use in Dashboards"
+          initialIsOpen={true}
+        >
+          <EuiFlexGroup>
+            <EuiFlexItem>
+              <EuiCard
+                icon={<EuiIcon size="xxl" type="discoverApp" />}
+                title={'Query data'}
+                description="Query your data in Data Explorer or Observability Logs."
+                onClick={() => {}}
+              />
+            </EuiFlexItem>
+            <EuiFlexItem>
+              <EuiCard
+                icon={<EuiIcon size="xxl" type="bolt" />}
+                title={'Accelerate performance'}
+                description="Accelerate performance through OpenSearch indexing."
+                onClick={() => {}}
+              />
+            </EuiFlexItem>
+          </EuiFlexGroup>
+        </EuiAccordion>
+        <EuiSpacer />
+        {TabbedPage({
+          tabNames: [
+            ['manage', 'Manage connections'],
+            ['new', 'New connection'],
+          ],
+          header: <></>,
+        })}
+      </EuiPageBody>
     </EuiPage>
   );
 }
