--- public/components/flint/components/accelerate.scss
@@ -1,4 +1,0 @@
-.accelerateCheckableCard {
-    width: 450px;
-}
-
