--- public/components/flint/components/accelerate_types.ts
@@ -1,11 +1,0 @@
-/*
- * Copyright OpenSearch Contributors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-export interface AccelerateProps {
-  pplService: any;
-  isFlyout: boolean | undefined;
-}
-
-export type AccelerateHeaderProps = AccelerateProps;
